+++ tests/dft_calculations/test_polarized_pbe.py
@@ -30,7 +30,7 @@ def test_polarized(system):
     xc = 'pbe'
     guess = 'random'
     etol = 1e-6
-    min = {'sd': 4, 'auto': 23}
+    min = {'sd': 4, 'auto': 24}
 
     atom, X = read_xyz(str(file_path.joinpath(f'{system}.xyz')))
     atoms = Atoms(atom, X, a=a, ecut=ecut, Z='pbe', s=s)